moving custom embedding strategies to specialized folder
