+++ src/attention_smithy/numeric_embeddings/__init__.py
@@ -1,5 +1,9 @@
-from attention_smithy.numeric_embeddings.SinusoidalEmbedding import SinusoidalPositionEmbedding, SinusoidalCustomEmbedding
+from attention_smithy.numeric_embeddings.SinusoidalEmbedding import SinusoidalPositionEmbedding
+from attention_smithy.numeric_embeddings.ALiBiEmbedding import ALiBiPositionEmbedding
 from attention_smithy.numeric_embeddings.RotaryEmbedding import RotaryPositionEmbedding
-from attention_smithy.numeric_embeddings.ALiBiEmbedding import ALiBiPositionEmbedding, ALiBiCustomEmbedding
 from attention_smithy.numeric_embeddings.LearnedEmbedding import LearnedPositionEmbedding
 from attention_smithy.numeric_embeddings.NumericEmbeddingManager import NumericEmbeddingManager
+
+from attention_smithy.numeric_embeddings.specialized.SinusoidalCustomEmbedding import SinusoidalCustomEmbedding
+from attention_smithy.numeric_embeddings.specialized.ALiBiCustomEmbedding import ALiBiCustomEmbedding
+
